refactor: move `ProjectSerializer`, `ProjectThumbnailSerializer` and `ProjectSeedSerializer` from `core` to `project` app
